requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
beautifulsoup4==4.14.2
boto3==1.40.57
botocore==1.40.57
certifi==2025.10.5
charset-normalizer==3.4.4
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
instaloader==4.14.2
jiter==0.11.1
jmespath==1.0.1
openai==2.6.0
packaging==25.0
playwright==1.55.0
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic_core==2.41.4
pyee==13.0.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
whitenoise==6.11.0
